src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceByteTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SingleReferenceByteTextParser extends SingleTextParser {

    public SingleReferenceByteTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Byte.parseByte(element.getText());
    }

}
