src/test/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceByteTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SingleReferenceByteTextParserTest {

    @Test
    public void constructorTest() {
        SingleReferenceByteTextParser parser = new SingleReferenceByteTextParser();
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
        SingleReferenceByteTextParser parser = new SingleReferenceByteTextParser();

        assertEquals((byte)2, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "2";
            }
        }));
    }

}
